requirements.txt
# ============================================================================
# Crypto-Stock Platform - Python Dependencies
# ============================================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & HTTP
asyncio==3.4.3
aiohttp==3.9.1
httpx==0.26.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Redis
redis==5.0.1
hiredis==2.3.2

# Exchange APIs
python-binance==1.0.19
alpaca-trade-api==3.1.1
yfinance==0.2.35

# Technical Analysis
TA-Lib==0.4.28
pandas==2.1.4
numpy==1.26.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Logging
loguru==0.7.2

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
watchdog==4.0.0

# Data Processing
pytz==2023.3
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
tqdm==4.66.1
click==8.1.7

# CORS
python-cors==1.0.0

# WebSocket
websockets==12.0

# Email (Optional - for alerts)
aiosmtplib==3.0.1

# Cloud Storage (Optional - for backups)
boto3==1.34.27
google-cloud-storage==2.14.0

# Load Testing (Development)
locust==2.20.0
